Connectors: rename parsers imports
